Name changes and linter fix
