+++ handlers/APIHandler.go
@@ -1,91 +1,56 @@
 package handlers
 
 import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
 	"net/http"
-	"os"
 	"v9_deployment_manager/database"
 	"v9_deployment_manager/deployment"
 	"v9_deployment_manager/log"
 	"v9_deployment_manager/worker"
-
-	"github.com/hjaensch7/webhooks/github"
 )
 
-type PushHandler struct {
-	actionManager *deployment.ActionManager
+type DeploymentIntentionHandler struct {
 	driver        *database.Driver
+	actionManager *deployment.ActionManager
 }
 
-func NewPushHandler(actionManager *deployment.ActionManager, driver *database.Driver) *PushHandler {
-	handler := PushHandler{
+type DeploymentIntentionBody struct {
+	ID                     worker.ComponentPath `json:"id"`
+	NewDeploymentIntention string               `json:"new_deployment_intention"`
+}
+
+func NewDeploymentIntentionHandler(actionManager *deployment.ActionManager, driver *database.Driver) *DeploymentIntentionHandler {
+	return &DeploymentIntentionHandler{
 		actionManager: actionManager,
 		driver:        driver,
 	}
-
-	return &handler
 }
 
-func (h *PushHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// TODO: Ensure we only deploy from master
-
-	// Load secret from env
-	secret, exists := os.LookupEnv("GITHUB_SECRET")
-	if !exists {
-		log.Error.Println("Failed to find Github secret")
+func (h *DeploymentIntentionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	// Parse Body
+	body, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Error.Println("Error reading body", err)
+	}
+	log.Info.Println(string(body))
+	var p DeploymentIntentionBody
+	err = json.Unmarshal(body, &p)
+	if err != nil {
+		log.Error.Println("Failed to unmarshal body", err)
 		return
 	}
-	// Setup github webhook
-	hook, githubErr := github.New(github.Options.Secret(secret))
-	if githubErr != nil {
-		log.Error.Println("github.New Error:", githubErr)
-	}
-	// Parse push event or installation event from webhook
-	// Note: integration events from github are ignored
-	payload, err := hook.Parse(r, github.PushEvent, github.InstallationEvent, github.InstallationRepositoriesEvent)
+	log.Info.Println(p.ID, p.NewDeploymentIntention)
+	// Update Database
+	err = h.driver.SetDeploymentIntention(p.ID, p.NewDeploymentIntention)
 	if err != nil {
-		log.Error.Println("github payload parse error:", err)
+		log.Error.Println("Failed to update status on database", err)
 		return
 	}
-	// Declare repo info vars
-	var user string
-	var repo string
-	var hash = "HEAD"
-	// Send to Installation Handler if needed
-	switch payload := payload.(type) {
-	case github.InstallationPayload:
-		log.Info.Println("Received Github App Installation Event...")
-		user = payload.Installation.Account.Login
-		for _, repo := range payload.Repositories {
-			compID := worker.ComponentID{User: user, Repo: repo.Name, Hash: hash}
-			h.processComponentEvent(compID)
-		}
-	case github.InstallationRepositoriesPayload:
-		log.Info.Println("Received Github InstallationRepositories Event...")
-		user = payload.Installation.Account.Login
-		for _, repo := range payload.RepositoriesAdded {
-			compID := worker.ComponentID{User: user, Repo: repo.Name, Hash: hash}
-			h.processComponentEvent(compID)
-		}
-	default:
-		parsedPayload := payload.(github.PushPayload)
-		user = parsedPayload.Repository.Owner.Login
-		repo = parsedPayload.Repository.Name
-		hash = parsedPayload.HeadCommit.ID
+	// Notify Action Manager
+	h.actionManager.NotifyComponentStateChanged()
+	// Send Response
 
-		compID := worker.ComponentID{User: user, Repo: repo, Hash: hash}
-		h.processComponentEvent(compID)
-	}
-}
-
-func (h *PushHandler) processComponentEvent(compID worker.ComponentID) {
-	// We want to ensure that we have the database stuff for this user built up
-	cID, err := h.driver.FindComponentID(&compID)
-	if err != nil {
-		log.Info.Println("Error finding database component id", err)
-	} else {
-		log.Info.Println("Received an event about component", compID, "| id =", cID)
-	}
-
-	// Then tell the action manager about the event
-	h.actionManager.UpdateComponentHash(compID)
+	fmt.Fprintf(w, "10/4 Good Buddy")
 }
