+++ main.go
@@ -72,7 +72,7 @@ func main() {
 	actionManager.NotifyComponentStateChanged()
 
 	http.Handle("/payload", handlers.NewPushHandler(actionManager, driver))
-	http.Handle("/api", handlers.NewAPIHandler(actionManager, driver))
+	http.Handle("/api/set_deployment_intention", handlers.NewDeploymentIntentionHandler(actionManager, driver))
 	log.Info.Println("Starting Server...")
 	err := http.ListenAndServe(CIPort, nil)
 	if err != nil {
